Nova: Switch mks console to Nginx
We have to switch our console over to Nginx, because we switch from the
`ticketType` "mks" to "webmks". The new ticket for the remote console
provides a WebSocket endpoint instead of a socket. This is not supported
by Nova's console code.

Since we already have token validation code for `shellinabox`, we re-use
that code and restructure it a little for the mks console. We do not use
`memcached` caching in the mks case, because mks does only a single
request and caching would just add overhead.

There's more potential for optimization which we plan to do in a
follow-up. For example, all cells currently query all DBs instead of
only their local one and the pods are not cell-az local. There's
probably also potential for more templating instead of duplicating
everything between shellinabox and mks.
