+++ openstack/nova/ci/test-values.yaml
@@ -19,6 +19,7 @@ global:
 
 imageVersion: myTag
 imageVersionVspc: vspc-latest-tag
+imageVersionConsoleMks: console-mks-tag
 
 mariadb:
   root_password: rootroot
