+++ openstack/nova/console/mks/10_default_server.conf
@@ -1,0 +1,56 @@
+lua_shared_dict token_cache 1m;
+
+map $http_upgrade $connection_upgrade {
+    default upgrade;
+    '' close;
+}
+
+server {
+  listen 6090 default_server;
+  error_log /dev/stderr debug;
+  resolver local=on;
+  default_type                application/json;
+  set $upstream               "";
+  proxy_buffer_size 128k;
+  proxy_buffers 4 256k;
+  proxy_busy_buffers_size 256k;
+
+  proxy_ssl_server_name       on;
+
+  proxy_connect_timeout       10;
+  proxy_send_timeout          30;
+  proxy_read_timeout          30;
+  proxy_http_version          1.1;
+  recursive_error_pages       on;
+
+  # Those options are in place to allow the iframe in elektra
+  add_header "X-Frame-Options"  "";     # Undo what is set in nginx.conf
+  proxy_hide_header "X-Frame-Options";  # Remove it from the upstream server
+
+  root /app/novnc;
+
+  location = /status {
+    access_log off;
+    content_by_lua_block {
+      ngx.say("Okay")
+    }
+  }
+
+  location = /websockify {
+    rewrite_by_lua_block {
+      local revproxy = require "revproxy"
+      ngx.var.upstream = revproxy.process()
+    }
+    # Support websockets
+    proxy_set_header Upgrade $http_upgrade;
+    proxy_set_header Connection $connection_upgrade;
+    proxy_pass $upstream;
+  }
+
+  location ~ ^/$ {
+    if ($http_user_agent ~ "^http-keepalive-monitor/") {
+      return 301 /status;
+    }
+  }
+
+}
