+++ openstack/nova/console/shellinabox/10_default_server.conf
