+++ openstack/nova/templates/console-mks-configmap.yaml
@@ -1,0 +1,21 @@
+{{ $envAll := . }}
+{{- $cells := dict "cell1" false }}
+{{- if .Values.cell2.enabled }}
+  {{ $_ := set $cells .Values.cell2.name true }}
+{{- end }}
+{{- range $cell_name, $is_cell2 := $cells }}
+{{- with $envAll }}
+---
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: nova-console-{{ $cell_name }}-mks
+  labels:
+    system: openstack
+    type: configuration
+    component: nova
+data:
+{{ (.Files.Glob "console/mks/*").AsConfig | indent 2 }}
+{{ (.Files.Glob "console/common/*").AsConfig | indent 2 }}
+{{- end }}
+{{- end }}
