+++ openstack/nova/templates/console-mks-secret.yaml
@@ -1,0 +1,22 @@
+{{ $envAll := . }}
+{{- $cells := dict "cell1" false }}
+{{- if .Values.cell2.enabled }}
+  {{ $_ := set $cells .Values.cell2.name true }}
+{{- end }}
+{{- range $cell_name, $is_cell2 := $cells }}
+{{- with $envAll }}
+---
+apiVersion: v1
+kind: Secret
+metadata:
+  name: nova-console-{{ $cell_name }}-mks
+  labels:
+    system: openstack
+    type: nova-console
+    component: nova
+type: Opaque
+stringData:
+  config.lua: |
+{{ include "nova.etc_config_lua" . | indent 4 }}
+{{- end }}
+{{- end }}
