+++ openstack/nova/templates/console-shellinabox-configmap.yaml
@@ -7,4 +7,5 @@ metadata:
     type: configuration
     component: nova
 data:
-{{ (.Files.Glob "shellinabox/*").AsConfig | indent 2 }}
+{{ (.Files.Glob "console/shellinabox/*").AsConfig | indent 2 }}
+{{ (.Files.Glob "console/common/*").AsConfig | indent 2 }}
