+++ openstack/nova/values.yaml
@@ -240,7 +240,7 @@ consoles:
   mks:
     enabled: true
     portInternal: '6090'
-    args: '--web /usr/local/noVNC-mks --verbose'
+    custom_deployment: true
 
 scheduler:
   driver: "filter_scheduler"
